requirements.txt
defusedxml
diff-match-patch
Django==3.2.*
django-admin-shortcuts
django-appconf
django-axes
django-compat
django-environ
django-filter
django-import-export
django-markdown
Django-Select2
djangocms-admin-style
djangorestframework
et-xmlfile
jdcal
lxml
odfpy
openpyxl
psycopg2-binary
PyYAML
tablib
unicodecsv
xlrd
xlwt
FreeSWITCH-ESL-Python